fix: 修复用户信息丢失
