+++ README.md
@@ -47,7 +47,7 @@ main分支为微服务版本，持续开发中
 - [ ] 消息模块引入大语言模型√ 每日定时做视频推荐（定时任务怎么写？）
 - [ ] 增加视频总结和关键词提取功能（大模型）AI视频推荐助手Agent落地实践
 - [ ] token的续期 双token？？
-- [ ] 项目快速部署和运维 K8s CICD体系
+- [ ] 项目快速部署和运维 github action√ K8s
 - [ ] 接入视频推荐算法（Gorse），对用户，视频，画像刻画
 - [ ] 视频格式大小校验√ 评论敏感词过滤，视频水印生成（FFmpeg）
 - [ ] redis一主两从哨兵模式 MySQL双主互从+Keepalived（redis和MySQL集群引入）
